refactor validation service architecture somewhat
